Change the way a received Webhook is converted to JSON
Reason: received webhook payload, if it contained a ListObject,
would be parsed as one thing and then when retrieved with stripe.Event,
it would look a different way.  This resulted in all webhook received being marked invalid.

Webhook view receives a text payload, to be interpreted as JSON. In the old way, this string is parsed with json.loads and then the resulting object is saved in JSONField.

In the new way, the text payload is given immediately to stripe.Event to be parsed into an Event object which is then serialized again to be stored in the JSONField. In this way Event data winds up in exactly the same shape for validated as for original.